app/api/pyproject.toml
[project]
name = "techpubsai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "techpubs-core",
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
    "azure-storage-queue>=12.9.0",
    "pydantic-settings>=2.2.0",
    "uvicorn>=0.30.0",
    "psycopg2-binary>=2.9.9",
]

[tool.uv.sources]
techpubs-core = { workspace = true }
